Merges main branch (Ekubo events, Curve StableSwap-NG/TwoCrypto-NG,
startup snapshot hydration, slot0 override mechanism, pool_creations)
into feature/logoperate-decoding (Fluid decoder, LogOperate events,
storage-based reorg handling, FluidPoolConfig cache).

All protocol enum/match sites updated with combined variants:
UniswapV2, V3, V4, Ekubo, CurveStable, CurveTwoCrypto, Fluid.

Key merge decisions:
- Startup snapshot loop from main + Fluid config resolution on snapshot
- Subscriber loop with Fluid config resolution on whitelist updates
- Both slot0 overrides (main) and Fluid orphan decode (logoperate)
  in reorg/revert paths
- Diagnostic tests use exhaustive pattern matching with _ => false fallback

65 lib tests pass + 3 ignored. Compiles clean (15 warnings).

+++ src/pool_creations/db.rs
@@ -9,7 +9,7 @@ pub struct PoolRow {
     pub asset0: String,  // lowercase hex
     pub asset1: String,  // lowercase hex
     pub creation_block: u64,
-    pub fee: Option<i32>,
+    pub fee: Option<i64>,
     pub tick_spacing: Option<i32>,
     pub additional_data: Option<serde_json::Value>,
 }
